Fix infinite loop in pcntl_sigprocmask
Summary: This code was relying on `sigismember` to tell you when you hit
an invalid signal, which it does on Linux but not OS X (it just says
invalid signals are not members of the set there). We can just cut off
when we hit the max number of signals, which POSIX helpfully specifies.
For comparison, this sort of loop is basically what PHP7 does, though
they check if `SIGRTMAX` is bigger than `NSIG`, which reading docs for
`NSIG` shouldn't be necessary, and is defined properly on Linux at
least.

Reviewed By: @fredemmott

Differential Revision: D2241420