Don't always regenerate systemlib.php in CMake build
Summary: This is one thing (though not the only one) standing in the way
of `make` on a built tree being a no-op: we are currently always
regenerating `systemlib.php`, which causes a re-link of the hhvm binary.
The reason for this is due to the usage of `add_custom_target`, which is
always considered out-of-date. Instead, we should use
`add_custom_command`, which tracks dependency information. However,
since lol CMake, depending on a custom command cross-directory doesn't
work AFAICT, so we still need a custom target that depends on that
custom command -- the custom target will always still get re-built, but
since all that does is call the custom command, it will be an effective
no-op. Whew.

Reviewed By: @fredemmott

Differential Revision: D2242785