Made most of the extensions optional.
Summary: This adds configs for all but the gd and imap extensions,
allowing most of them (the ones not explicitly marked as REQUIRED)
to be disabled via the ENABLE_EXTENSION_* variables.
This also adds configs for the extensions that were originally
in the root directory.

This supersedes PR #5413.
Closes #5571

Reviewed By: @jwatzman

Differential Revision: D2216843

Pulled By: @sgolemon