Remove size class from header to make some room
Summary: Remove the 6-bit size class index in header so that we can use the bits for
the gc project.  We will need to recalculate the size class index when
releasing a StringData. It will probably be a small regression.

It still seems the right thing to do, as the gc project will potentially enable
much better wins with these bits.  Hopefully we can improve the CapCode encoding
so that it is easy to extract the size class to achieve a similar result (not
trivail, though).

Reviewed By: @edwinsmith

Differential Revision: D2242897