Make NEVER_INLINE usages MSVC compatible
Summary: Because MSVC doesn't like `__declspec()` attributes after the arguments.
Hopefully I got all the uses in this PR, but I didn't check every single use of the attribute, as the vast majority are already placed before the function name.

This will conflict with #5632 (D41565) due to a spot in the asio extension where both `ATTRIBUTE_NORETURN` and `NEVER_INLINE` attributes needed to be moved.
Closes https://github.com/facebook/hhvm/pull/5678

Reviewed By: @​svcscm

Differential Revision: D2237035

Pulled By: @sgolemon