Remove the global LibIConv requirement
LibIConv is currently required globally, but only `ext_iconv` and `ext_gd` use it. `ext_iconv` needed no changes, as it already uses the new extension mechanism, but `ext_gd` doesn't. Because of this, the iconv requirement was inlined in `ext_gd`.
This also implements LibIConv in the extension config mechansim.