Remove the global FastLZ requirement
The only thing that uses fastlz currently is `ext_memcached`, if we already aren't building that, we don't need it.
This one is a bit more fun than the rest, because fastlz currently falls back to a bundled version if the system doesn't have it.
This also adds the ability to force the use of the bundled version via `HHVM_USE_BUNDLED_FASTLZ`. This ability will not however work until the way the 3rd party repo handles it is also adjusted.

Now, obviously I can't tell CMake to link fastlz against itself, so I use a generator expression to enable the library in the link string only if the target is not named "fastlz". This works fine, however, some versions of CMake complain that I'm still trying to make fastlz dependent upon itself. To get CMake to be quiet about this, we set policy CMP0038 to the OLD behavior, which is to simply ignore that dependency. (the NEW behavior is to error)

For this to work fully, changes to the third-party repo will also be needed (enable fastlz only if `HHVM_USE_BUNDLED_FASTLZ` is both defined and `true`, rather than checking if `FASTLZ_INCLUDE_DIR` is valid), but it still works without any changes to it.