Port the JIT translator helpers to MSVC
Because Microsoft has never heard of AT&T syntax.

This also enables the ASM_MASM langauge for the HHVM project. CMake will ignore whichever language, ASM or ASM_MASM, the current platform can't support, at least, that's how it handles ASM with the MSVC target.
I've checked this against the dissassembly of the AT&T syntax version, and they are identical, except for the difference in which registers are used due to the Windows ABI.