Minor tweaks in string hash functions
Summary: Minor adjustment in the asm implementation of hashHelper.  Basically
combine `neg` and `jz` to save a `test`, except that we do a `sub` from
memory here.  We then use the space saved to avoid doing a bulky 10-byte
movabs in the loop.

The unaligned case-insensitive hash function is changed to use the aligned
implmentation if the string is aligned.  This should avoid the byte-by-byte
read for the tail in most cases.  Thanks @​sathvik for identifying the
bottleneck in open-source builds.

Reviewed By: @edwinsmith

Differential Revision: D2216933