Remove the global check for libpam
Summary: The PAM headers are not used anywhere in HHVM or any of it's dependencies as far as I can tell, and no indication is even made that we have libpam in the first place.
This also removes the FindLibpam.cmake file, as it's not needed if we aren't doing this check.
Closes https://github.com/facebook/hhvm/pull/5728

Reviewed By: @fredemmott

Differential Revision: D2254996

Pulled By: @sgolemon