Remove the global MCrypt requirement
Summary: Because `ext_mcrypt` is the only thing that needs it, and the extension config mechanism handles adding the includes and libraries.
This also moves the mcrypt section in the extension config mechanism to the correct spot in the alphabet. I have no idea how it ended up down there, but it shouldn't have been.
Closes https://github.com/facebook/hhvm/pull/5725

Reviewed By: @fredemmott

Differential Revision: D2254990

Pulled By: @sgolemon