Expose client perf stats and timings with Results
Summary: This exposes the data about the perf stats the client had in the moment
the operation finished. This way PHP side can log as it wants the
information and it will be used to identify if the errors were caused by
ASIO being overloaded or the IO Thread.
For the `AsyncMysqlConnection` we will have `lastActivityTime` that
contains the last time a successful query was made or it "connected".
Also, I added to it the AsyncMysqlConnectResult so the timing the
operation started and completed are accessible without noise of passing
values between threads (Going from async mysql IO thread to PHP thread).

Perf stats were added to all AsyncMysqlResult's, this way when an error
happens we have the state AsyncMysqlClient was.

Reviewed By: @​anca

Differential Revision: D2253236