Fix some issues with extract
Summary: extract can cause destructors to be called, which
can have arbitrary side effects including clobbering the
array that was passed into extract in the first place. Doing
so would result in a call to lvalAt on something that might
not even be an array. The new test case would assert in dbg
builds, and crash randomly in opt builds.

In addition, builtins are not allowed to mutate their parameters,
because the jit assumes it can decRef the SSATmp that it originally
passed in on return from the call - so don't pass a reference to the
original array unless it was boxed. This showed up in D2235427,
because after that we can pass in an unboxed array with a refcount
of one.

Reviewed By: @swtaarrs

Differential Revision: D2264520