ext_hh_client is required
Summary: If it isn't available, the auto-typecheck logic will try to
call a function that doesn't exist. This is largely academic since it
has no dependencies outside of those of the typechecker itself (which we
require you to build nowadays), but just in case someone tries to turn
it off explicitly, let's be clear.

Reviewed By: @paulbiss

Differential Revision: D2260956