Fix resource generation on Windows
Summary: I'm not entirely sure why it wasn't working before, but I believe it was caused by an issue with CMake itself. Hopefully this still works on the previous environement as well, but I know it works for me now on Windows with MSVC.
This also makes it work properly with out-of-tree builds.
Closes https://github.com/facebook/hhvm/pull/5659

Reviewed By: @fredemmott

Differential Revision: D2235911

Pulled By: @sgolemon