Remove the global LibBZip2 dependency
Summary: Because the only thing in the repo that has the string `bzlib.h` in it, apart from pcre (in the third-party repo), which finds it on it's own, is `ext_bz2`.
This adds support to the extension config mechanism for handling LibBZip2.
This also renames the dependency in the configuration from `bz2` to `bzip2`, to better reflect the name of the actual library.
Closes https://github.com/facebook/hhvm/pull/5721

Reviewed By: @fredemmott

Differential Revision: D2254983

Pulled By: @sgolemon