Add counters for exceptions thrown in error handler
Summary: When a request OOMs or times out (or raises some other
exception) HHVM runs the user error handler (a PHP function) to
log/handle the exception. If called while in PSP the error handler also
runs the rest of the end-of-request logging/cleanup (like populating
data that goes into dynolog) that would happen if PSP hadn't been
killed:

https://fburl.com/134901587

If we OOM (or presumably time out/etc) in the error handler again we'll
just eat the exception and don't do any of the cleanup. The dyno line we
log in this case is almost useless:

  1437568244 500 1066235 871929 72066 71990 - - - - - "/api/graph/server.php|/insights" graph.facebook.com - - - 109 - - - - - - "Internal Server Error" - - "-" - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - 3521422277 218737394 "-" "-" "-" - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - 1 1 0 0 - - - - - -

This diff adds some counters for when we throw from the error handler.
It doesn't fix any of the underlying problems that cause this to happen
but as least it'll let us see what fraction of requests are not getting
logged correctly. I want to improve this situation so a these stats will
help us tell if we're getting better or worse.

Reviewed By: @markw65

Differential Revision: D2268370