Add support for static props and methods to experimental emitter
Summary: Add support for static properties and methods to the experimental emitter. This involves tracking the name of the self and parent class in order to emit classrefs and refactoring lval emitting to be able to put things on the stack in the bizarre order that member instructions need things when accessing static property bases.

Depends on D2262003.

Reviewed By: @int3

Differential Revision: D2262066