Fix usage of undefined behavior in TypedValue
Summary: Often we want to manipulate the ref-counts of an object represented by
a TypedValue, but we don't statically know the actual type of the object. The
common idiom for this is to choose a type, and perform the ref-count
manipulation using the pointer of that type. IE: m_data.pstr->incRefCount(). We
rely on the fact that all of the types have the same ref-count manipulation
functions, and therefore will work even if the actual type doesn't match the
pointer type we choose.

However, this is a violation of strict type aliasing, as we are manipulating an
object via pointers of unrelated types. The compiler is free to assume that such
accesses are to different objects and optimize accordingly. Most of the time we
get away with it, as the manipulations are far enough apart, or in different
compilation units. However, there's been at least one case observed where the
compiler miscompiles an inc-ref done inside a TypedValue (pointing to an
ObjectData), and a following dec-ref done on the same object via a ObjectData
pointer. The order of the inc-ref and dec-ref are changed so that the ObjectData
can be freed early.

To avoid this, introduce a new macro called TV_GENERIC_DISPATCH(), which takes a
TypedValue, and the name of the function to invoke on contained pointer.  The
num TypedValue member (which contains the pointer value) is incremented by
FAST_REFCOUNT_OFFSET, cast to a pointer to RefCount, then passed to one of a set
of generic ref-count manipulation functions. This is effectively what we are
doing now, but by avoiding using any of the pointer types, we do not violate
strict type aliasing.

Not only does this fix the issue (the miscompilations are gone), but in some
situations even better code is compiled. The compiler has been observed to elide
needless inc-ref/dec-ref pairs in a few situations.

Also remove a few calls to tvIncRef in tv-helpers.h where we statically know
what the type is.

Reviewed By: @markw65

Differential Revision: D2261012