Fix HH\Asio\curl_exec by hostname where cURL is built without ares
Summary: ares is an asynchronous DNS library, which provides a socket that
curl (and asio) can select() on. On systems without it (eg Ubuntu), standard
libc resolution is used, and we dont' have a socket, so curl_multi_await()
returns -1 until DNS resolution is finished.

This also affects curl_multi_select() in PHP5 but not HHVM as we use
curl_multi_wait() instead of curl_multi_fdset(), but we need to use
curl_multi_fdset() for async.

fixes facebook/hhvm#5384

Reviewed By: @sgolemon

Differential Revision: D2274514