Add support to HPHPSetup for generation targets supporting multiple configurations
Previously they would be forcing either debug or release, which isn't needed when the generation target, such as MSVC or XCode, support multiple configuration types.
CMake also already adds the `DEBUG` and `NDEBUG` defines for the appropriate configurations to begin with, so we simply don't add any definitions for multi-config targets.