add getCounter() function in fb303 server handler to support retrieval of specific counter by key
Summary: This diff adds getCounter() function in Fb303Handler class. Thrift client thus can use this function to retrieve the counter value by key name. This reduces the time spent in thrift call from a few thousands microseconds to a few hundred microseconds. (Before this, to get the value for a specific counter, one needs to first call getCounters() to get the all the key-value pairs for counter, timeseries and histogram and then look up the returned map to find the value of a key.)

Reviewed By: @markw65

Differential Revision: D2225020