Make accessing an abstract type constant an error
Summary: Currently the type checker will allow you to access a type constant even if we know that the type constant is abstract. For example:

```
abstract class C {
  abstract const type T;
  abstract protected static function getTImpl(): this::T;

  public static function getT(): this::T {
    return static::getTImpl();
  }
}
```

The type ##C::T## is currently allowed but it is a type that can never be enforced by the runtime. Since we do want to have type constants enforced by the runtime at some point let's make this an error in the type checker. It's questionable if this type is useful at all so don't think we are losing much for doing this.

Another wrinkle is the type ##this::T## for the method ##getT##. The method ##getT## is a valid declaration, but invoking ##C::getT()## should be an error because ##this::T## will result in the type ##C::T## after we resolve the ##this## type. This means that the error will occur when we localize the type ##C::getT##. This is a bit surprising, but actually captures a real error case, in that invoking ##C::getT()## will always produce an error at runtime. This problem exists with the ##this## in general and static methods, but for now I only generate an error when this occurs for type constants.

I also improved the errors we report when we fail to expand a type constant. Without these improved errors it is impossible to understand what exactly needs to be fixed when this error is reported so I included it in this diff.

Reviewed By: @jwatzman

Differential Revision: D2276750