Fix resource generation on Windows
Because RC doesn't know that double quotes aren't part of the name of a resource.

While my previous fix did get it compiling, it was flawed in-that it caused double quotes to be included as part of the actual name of the resource. This takes a different approach to solving the problem at hand, and replaces the `.` which was causing the problem, with an `_` instead.
This also properly assigns the resources to the neutral culture, rather than whatever the culture of the build machine is.
This also fixes a minor typo in the name of the `MINGW` variable in the `if` statement controlling resource generation in the `CMakeLists` file for HHVM.