Remove an unneeded HANDLE being opened on Windows
`FindResource`, `LoadResource`, and `SizeofResource` all default to the module that was started up, the executable, so don't open a `HANDLE` explicitly, especially not one that's never closed.
This also handles some minor formatting changes, and finally, it moves the `#include`s for Windows around so that Windows doesn't try to include the elf headers.