Support detection of the home directory on Windows when creating the global repo
Because the `HOME` environment variable is not set on Windows, and there is no password database to lookup the home directory in on Windows.
The order and which checks are used should be the same as what Qt uses to determine the home directory. If it isn't, then I blame Stack Overflow.