Reduce HHVM's dependence on the symlinks in third-party
Summary: Because these symlinks don't translate properly on Windows (git just checks them out as plain text files), and would have to be manually re-created. It's better to simply not use the symlinks in the first place and avoid the hassle.

`libmbfl/filters` was also misspelled as `libmbfl/filter`, so this fixes that as well.
Closes https://github.com/facebook/hhvm/pull/5660

Reviewed By: @sgolemon

Differential Revision: D2235912

Pulled By: @sgolemon