Fix ref-count handling for IncDecl bytecode
Summary: When emitting IR for the IncDecL bytecode, the proper number of inc-refs are not
emitted. An IncRef is emitted to account for the value being pushed onto the
stack, and a DecRef is emitted to account for the old value in the local being
overwritten. However, an IncRef is not emitted for the new result being stored
in the local. As a result of this, a value can be released early, causing memory
corruption. The solution is simple, ensure the value is inc-ref'd when stored as a local.

This was originally found while running -vEval.EnableGC=1 on the
hphp/test/zend/good/Zend/tests/decrement_001_64bit.php unit test. Decrementing
an object is a no-op, but the object was released early when it was popped off
the stack.

Reviewed By: @alexmalyshev

Differential Revision: D2283851