Give hh_emitter an option to take its input code from stdin
Summary: Add a --stdin option to hh_emitter that causes it to read code to compile from stdin. It still takes a command line argument for the filename, which it wil use for debug information and the like. The rationale is to allow it to be driven directly by hhvm while directly using its source of truth for file contents.

Reviewed By: @int3

Differential Revision: D2285337