Fix platform quirk of get_sockaddr
Summary: So, not 100% sure what is going on here, but here's my guess.
We use our function `get_sockaddr` to parse the result of the stdlib
function `getpeername` (and others, but that's the important one here).
For unnamed `AF_UNIX` sockets, on Linux, that function will return
a length that is too small and garbage where the string should be; we
specifically look for this case, and if it happens, skip reading the
data and setting the address, so it comes back as null. However, on OS
X, it looks like the stdlib explicitly returns an empty string, which
we'd dutifully write into this struct, giving a different result.

So, I had to make a judgement call which way to go here; it seems that
explicitly writing an empty string is better than just leaving the ref
unset, so I did that. This could of course go the other way if that
causes problems (though all the tests pass), by only setting the address
if `strlen > 0`.

FWIW, it looks like PHP7 doesn't deal with this at all, and reads
garbage? There may be other magic going on.

Reviewed By: @markw65

Differential Revision: D2185103