Properly handle the 'this' type in type constant constraints
Summary: For type constants whose as constraints contained a `this` type we were not
properly resolving the 'this' type to an expression dependent type. We did this
for the `this` type in the assigned type of a type constant, so I moved the
logic outside the pattern match so it applies to all cases.

Reviewed By: @jwatzman

Differential Revision: D2286736