Fix various issues with reference parameters
Summary: We assume that SSATmps can live across FCallBuiltin,
so we DecRef the passed in SSATmp, rather than reloading the
value from the stack after the call.

This is a problem for by-reference parameters, because they're
not required to actually be passed by reference.

This changes the api for VRefParam so you can't accidentally mutate
it, and then attempts to fix up all the fallout. There were a few
interesting bugs - eg sscanf manages to crash when not given a
reference even though its *never* called via FCallBuiltin (new test
case added).

Fixing things was mostly straightforward; but in some cases the fixes
are tricky, or the simple fix would force a copy-on-write, of an array
which we then access one element of, and discard. I've tried to avoid
cases like that whenever possible.

Reviewed By: @paulbiss

Differential Revision: D2290579