Move from folly::dynamic to Variant for ini settings
Summary: `folly::dynamic` was actually pretty good. But we require certain functionality
(e.g., maintaining insertion order of a map) that it could not easily provide.
`Variant` helps better provide what we need, and now that certain issues are not
issues any longer (e.g. lifetime) because we parse the ini map and just throw
it away, we can use it.

This reduces confusion in the code a bit and eliminates some code as well.

Reviewed By: @markw65

Differential Revision: D2050842