remove TypeStructure typedefs in hhvm hhi
Summary: In preparation for typechecker changes, add type_shape<T> as TypeStructure so that the runtime has knowledge of the type_shape type.

However, currently typedefs don't work in hhvm, hence the plan is to add these to www.hhi when the time is right or when the issue gets fixed.

Reviewed By: @paulbiss

Differential Revision: D2292376