Teach HHVM about Windows paths
This adds a `FileUtil::isAbsolutePath` function that is used in-place of checks that were being done to see if a path was an absolute path.
This also replaces a couple of places that were doing these checks with calls to this function. This does not do this everywhere, and instead has only been done where it was causing errors when including files on Windows.
`FileUtil::dirname_helper` and `FileUtil::canonicalize` have both been taught that a backslash is also a directory separator.