Refactor vasm toplevel emit loop
Summary: The toplevel emitter loops for both the x64 and the ARM vasm backends
were nearly identical.  Refactor this loop as a routine which takes a
single-instruction emitter class as a template parameter, and which
implements the mechanism for two-phase emit (emit all code, then process
patches).

The Vemit contract was intentionally designed to avoid tedious codemod
on the arch::Vgen method implementations.

Reviewed By: @edwinsmith

Differential Revision: D2279048