Improve hphpd's variable command.
Summary: Each variable's value gets fully serialized and sent to the client using the
VariableSerializer::DebuggerSerialize kind.  Then the client unserializes it,
and serializes it again but now with the DebuggerDump kind.  Now it realizes
that it might not want to dump this potentially huge value, so it helpfully
truncates it....after serializing it twice already.

Change the client to send over the user's serialization limit with the command,
and have the server do the truncation.  Furthermore, have the server serialize
with DebuggerDump, and have the client just dump that result.

Note, the previous behavior of hphpd was to report the number of characters
remaining when printing a long value and stopping early.  Since we no longer
serialize the entire value in the client, we can't know how many characters are
left, so the prompt now just says "There are more characters" instead of "There
are 5000000 more characters".

This is the majority of "wire errors" and socket errors I've seen users hit in
hphpd.  It usually happens when a user tries to do "=EXPR;" in hphpd, and
somewhere in EXPR lies a value that we can't round-trip through our serializer,
most commonly because it has an object with a NativeData somewhere in it.
Furthermore this should help performance of printing variables, as now we won't
be serializing them in entirety unless the user asks us to.

Reviewed By: @paulbiss

Differential Revision: D2278646