Fix session.save_path
Summary: It can be a relative path, so we need to set it
at the start of each request.

Even for absolute paths, the normal restore mechanism can
fail, eg due to eg a call to open_basedir making the
default path illegal at the time we try to restore it
(just because someone called

  ini_set("open_basedir", "/foo/bar/baz");

on this request, doesn't mean they'll do so again on the
next one).

Reviewed By: @JoelMarcey

Differential Revision: D2301167