Fixed the WDDX extension's comparison with a type-expressed-as-string
Not sure why this extension doesn't just compare on the real type, but issues
string compares with a stringyfied version of the type name. Then again, I
doubt anybody is really using WDDX?

I left the original comparison with NULL in here as well. It's not needed for
any tests to pass, but I thought it to be prudent to keep. Happy to take it out
if that's what you'd like.