resolve aliases in TypeStructure modified
Summary: As discussed in t7888530, fully resolve type aliases might cause fatal in the runtime due to the inconsistency between hack and hhvm on non-hoistable classes. Hence we postpone resolution of type aliases until request time (via ReflectionTypeAlias). For TypeStructure of type constants, we do not resolve the aliases unless they are used in type accesses.

The implications are the following:
--users must declare all relevant classes appeared in the type alias (or nested aliases) before calling ReflectionTypeAlias->getResolvedTypeStructure(), otherwise the method will fatal;
--for aliases in type constants that are not used as accesses, the user must use ReflectionTypeAlias to further resolve those. (D2304554 foregoes this assumption.)

Reviewed By: @paulbiss

Differential Revision: D2304258