Get naming/ to use Core
Summary: One of the nice properties of Core is that all potentially
exception-raising functions are named `*_exn`. The non-exn-suffixed
functions return an option type if necessary. In particular, List.hd
returns an option type. This caught one case that we weren't handling --
when the user types `assert()` without any arguments, we would crash.

Reviewed By: @dabek

Differential Revision: D2303725