Don't fail on large allocations in ReadOnlyArena
Summary: ReadOnlyArena used to assert that we never tried to allocate regions
bigger than m_chunkSize. So far this has caused nothing but operational
headaches, so I've modified it to support variable-sized chunks as
needed. There's still a configurable minimum chunk size since that can affect
how much space is wasted in the tails of chunks.

Reviewed By: @edwinsmith

Differential Revision: D2302310