Fix a couple of issues with ExitPlaceholders
Summary: LICM assumes the ExitPlaceholders are emitted before a block's guards,
which was originally true.  There doesn't seem to be any good reason
for emitting them after the guards, so move them back before the
guards, and change LICM to assert that it finds an ExitPlaceholder.

Also, ExitPlaceholders were being removed from the unit very early,
before LICM.  So delay removing these instructions as well.

With HHIRConstrictGuards, ExitPlaceholders were being inserted to
indicate the end of the guards in a tracelet, which is the purpose of
the EndGuards instruction.  So stop using ExitPlaceholders for that
purpose, and instead emit EndGuards in the desired program points.
This allows us to clean-up visitGuards a bit too.

Reviewed By: @swtaarrs

Differential Revision: D2308060