Eliminate arch-dependency for service request emission
Summary: This diff moves the work of emitting service requests into the main vasm
emit loop.  It also decouples the emission of the service request itself
(e.g., the smashable jump) from the service request stub, which obviates
a lot of edge-case logic that arises when we run out of stub space for a
service request that is emitted to frozen.

When the emit routine is called on a service request, we now only emit
the smashable jump, and register the need for a stub.  The vasm_emit()
loop then emits the stubs, and uses the existing patch point machinery
to trigger architecture-dependent patching of those jumps to point to
the stubs.

Since the inline portion of all service requests is essentially "emit a
smashable jump", once the code smashing API is architecture-independent,
we can kill the emit() rountines in place of lowering emits in
vasm_emit().

Reviewed By: @ottoni

Differential Revision: D2284158