Add runtime option to eagerly run GC
Summary: Add a new runtime option called "Eval.EagerGCProbability". If set to a value
greater than zero, a garbage collection will be eagerly run at every
allocation or deallocation with probability 1/N. Eval.EnableGC must be set to 1
for this option to do anything. To avoid overhead in production, this option is
only available in debug builds.

This is implemented by doing a random coin-flip with the specified probability
at every allocation or deallocation. If the coin-flip succeeds, a new surprise
flag is set. A garbage collection is then done at the next surprise flag
check. This is safer than attempting the garbage collection in the middle of the
allocation. In the future, the surprise check flag will be used when we need to
run a GC for real.

Right now enabling this option often causes asserts to be triggered, because the
heap isn't always in a fully consistent state even during the surprise flag
checks. This is being worked on.

Reviewed By: @edwinsmith

Differential Revision: D2315369