emit type information for function args and return values
Summary: Generate TypeConstraint information for function arguments and return values. This is kind of janky, mostly because TypeConstraint is kind of janky. We don't emit typechecks for parameters/return values yet; the value in having the annotations anyways is in being able to pass the checks that parent/child classes agree when combined with code not from the HHVM emitter.

Depends on D2310618.

Reviewed By: @int3

Differential Revision: D2310747