Fix a bug printing annotations in tc-print
Summary: There was a one-off bug which was causing the annotations to be
sometimes cut short.  It seems that the code was accounting for the
'\0' at the end of the string, but that character is inserted by
gzgets and not read from the file, and therefore shouldn't be
accounted for.

Reviewed By: @maksfb

Differential Revision: D2322229