Fix instrInputFlavor for member vectors
Summary: It thought S bases implied A on top with C below it. The C it was
thinking of is actually elsewhere in the eval stack, but it happened to be ok
because there was a C there anyway for other reasons. I simplified the function
a little bit so it only special-cases non-C flavors.

Reviewed By: @edwinsmith

Differential Revision: D2315417