Fix LICM's usage of ExitPlaceholders
Summary: LICM was reusing the exit block from the loop header's ExitPlaceholder
as the target of checks hoisted out of the loop.  That's not always
correct because that exit can use values defined in a DefLabel at the
loop header.

This diff fixes the problem by cloning the exit and renaming SSATmps
accordingly, so that it can be used before the loop.  This `preExit'
block is cached for the loop, and it can be reused by multiple hoisted
check instructions.

Reviewed By: @paulbiss

Differential Revision: D2315881