Stop using ActRec in DomDocument
Summary: This took some refactoring because DomDocument::load...()
depend on different behavior when they're called statically
and we made the decision to not support that in normal HNI methods
(hence the use of ActRec here).

I'm moving the patently ridiculous `if (!$this)` logic into PHP
where it's "properly" supported by the VM, and unifying the helpers
down to just XML versus HTML, with isFile as a flag.

I also disabled a dumb test which doesn't test anything useful.

Reviewed By: @paulbiss

Differential Revision: D2320987