Add specialized array comparison IR ops
Summary: Add new IR ops for array comparisons, named SameArr, NSameArr, EqArr,
NeqArr, GteArr, GtArr, LteArr, and LtArr. These ops have the same semantics as
the generic comparison ops with array operands. Teach the simplifer to convert
any comparisons involving array operands to the object-specific ones.

Since we always punted on non-equals array comparisons, there were some bugs
with array comparisons in the simplifier.

This is a step towards migrating over to specialized comparison ops completely,
removing the need for the generic ops. After this, only resource comparisons
remain. After that, the generic ops can be removed, and a lot of code cleanup
can be performed. The simplifer is getting a bit messy, but that's temporary.

Reviewed By: @edwinsmith

Differential Revision: D2311888