Implement the WANTED state for extensions
Currently they use the explicitly enabled state as their default, which not only makes the error message incorrect, but also makes it impossible to allow extensions that fail a `var*` or `os*` dependency to silently disable themselves without causing extensions that really have been forcefully enabled to not error when these dependencies fail.
This implements a WANTED state, and makes it the default. The WANTED state allows extensions that fail `os*` or `var*` dependencies to disable themselves cleanly, while still allowing these failures to trigger errors if the user has explicitly enabled an extension where this occurs.
This also adds a warning if you attempt to explicitly disable an extension that has been marked as 'REQUIRED'.
Lastly, this removes the `EXPLICIT` modifier on `ext_gmp`, because it isn't needed.