Improve scan function coverage. Hook up generated ResourceData scanners.
Summary: Improve scan function coverage.  Scan functions are now generated for all subclasses of "interesting" types and all classes that contain "interesting" types.

Add better handling for typedefs, templates, cloned and anonymous classes.
Add list of classes that are ignored during scan process.
Add more classes to hasScanMethod set (these are defined on the IMarker class).
Add gcContainerNames set for classes that hide uses of scanable objects with casts or other tricks.  Any use of one of these classes with a scanable template argument is considered scanable.

The generated scan functions for ResourceDatas will be hooked up if the generator is run in a sandbox.  Otherwise, scanning will fall back to the default.

Reviewed By: @edwinsmith

Differential Revision: D2053297