Make the CMake generated project files for HHVM easier to work with in Visual Studio
Summary: By default, they are flat, and every file shows up right next to each other, which is not very easy to manage. This can be changed by adding `source_group` calls, one per directory. To achieve this, I have added an `auto_source_group` function which accepts the root name, root directory, and the sources to group. It will then create source groups in such a way as to mimic the file tree on disk.
This also adds the headers to the compilation of various targets, because they will not show up in the IDE otherwise.
Closes https://github.com/facebook/hhvm/pull/5620

Reviewed By: @jwatzman

Differential Revision: D2233796