Allow instanceof on unconstrained generics
Summary: This makes `classname<T>` more useful.

Even if `T` turns out to be a primitive type or enum at runtime, it's
still sound as far as instanceof is concerned -- if something passes the
instanceof check, then we know T is a class; if the check fails, we
don't make any assumptions on what T is.

Note that what is *not* sound is calling `new` on an unconstrained
generic, but that's an existing issue.

Reviewed By: @dlreeves

Differential Revision: D2325706