Remove some verbose logs from filesystem extension
Remove some verbose logs from filesystem extension
A couple of functions are throwing some ugly errors (e.g. when the file
does not exist). This is a small incompatibility with PHP, as those functions
are not supposed to log those kind of errors, so I removed them.
I also removed some commented code which has been dead for a long time.

No test added/modified (seems that this extension is barely covered with tests).