Give BootTimer samples for warmup scripts a unique key
Summary: We run each warmup script twice. IIRC the second time is kind of a "happy bug", in that it helps jit the warmup scripts and results in better CPU numbers.

That's very nice, but adding the timings for both runs is misleading and makes any analysis worthless.

While here, also clean up the path leaving only the last component.

Reviewed By: @​markdrayton

Differential Revision: D2322459