Make references a unop and push them through to the Nast
Summary: Change references from their own ast node to being a sort of unop. This has causes them to be properly reflected in the Nast instead of being silently dropped. Since &$x acts like an lvalue in in foreach value bindings, typechecking needed to learn how to deal with that.
My main motivation here is exposing references to hh_emitter, but in general I think that more Nast fidelity for important things like references is a good idea.

Reviewed By: @int3

Differential Revision: D2324599