Change NULL to null for error reporting and internal type queries.
Summary: This addresses an incompatibility with PHP 5 for displaying a string version
for the null type.

PHP 5 returns "NULL" through gettype():
https://github.com/php/php-src/blob/master/ext/standard/type.c#L34

HHVM returns "NULL" through getDataTypeString():
https://github.com/facebook/hhvm/blob/master/hphp/runtime/base/type-string.cpp#L571

However, when asking in an extension for the type of a null value, or with
parameter parsing errors, PHP and HHVM where doing something different.

PHP returns "null":
https://github.com/php/php-src/blob/master/Zend/zend_API.c#L162

and HHVM reuses getDataTypeString() and hence uses "NULL":
https://github.com/facebook/hhvm/blob/master/hphp/runtime/base/type-string.cpp#L571

This patches changes getDataTypeString() to return "null" to be in line with
PHP 5, and has a special case before getDataTypeString() gets triggered with
gettype() to return "NULL".
Closes https://github.com/facebook/hhvm/pull/5828

Reviewed By: @jwatzman

Differential Revision: D2300810