Allow ini files to include other ini files
Summary: Currently with hdf we allow the inclusion of other hdf files within
an hdf file. This provides that support for ini.

Basically, if you have a line that has:

```
```

in your ini file, then the master ini string gets populated with its contents.

All this is done before the actual ini setting processing. And it maintains
order semantics as well (i.e., the contents of an included file is placed
exactly where the include was provided).

Reviewed By: @markw65

Differential Revision: D2144743