Allow setting LIBPTHREAD_LIBRARIES to a multi-config friendly value
Summary: Without this, if you've set `LIBPTHREAD_LIBRARIES` to a multi-config friendly value, such as this except from my CMake cache:
```
LIBPTHREAD_LIBRARIES:FILEPATH=optimized;F:/Libraries/libpthread/lib/libpthreadMT.lib;debug;F:/Libraries/libpthread/lib/libpthreadMTd.lib
```
CMake will error when it tries to do this, because it is incorrectly trying to substitute the value of this variable in the condition of the if statement.
This simply changes it to not substitute the entire value.
Closes https://github.com/facebook/hhvm/pull/5881

Reviewed By: @jwatzman

Differential Revision: D2328069