Use C++11 attributes rather than __attribute__
Because [GCC 4.8](https://gcc.gnu.org/gcc-4.8/cxx0x_status.html) supports them, as does MSVC, and they are much cleaner than the other syntax.
MSVC doesn't support the `[[gnu::]]` attributes, but that is just a warning, which we can disable.

This also adjusts a couple of uses of `__attribute__((__noinline__))` to `NEVER_INLINE`, to work correctly with MSVC.