Fix JIT array type specialization
Summary: A constant array may be regarded as a subtype of multiple
different specialized array types, e.g., a static packed array can
be a subtype of array-kind specialization Arr:PackedKind, or a
subtype of RAT specialization.

The current approach only look at array-kind specialization for
constants, and it can cause Bottom to be generated in load elimination
after AssertRATI.  This diff fixes it by looking at multiple
ways of specializing the array type.

Reviewed By: @mxw

Differential Revision: D2332258