Make the assembler compute hoistability
Summary: Make the assembler actually compute class hoistability instead of blithely marking everything MaybeHoistable.

The assembler is very "on-line" in that it basically parses straight into calls into *Emitter functions, without any intermediate data structures. This means that properly acting on the hoistability required breaking up some logic in UnitEmitter that depended on having the hoistability when the PreClassEmitter is created.

Reviewed By: @markw65

Differential Revision: D2312025